Enabled authentication repositories.  Fixed concurrency issue arising from access control
Access requests come in on thread A but are verified on thread B.
Additional HTTP requests come in on thread A while verification may not
have yet completed on thread B.

Solved this by queueing requests until access verification completes.
Once access verification completes the queue is drained and processed.

TODO: look into other areas that could suffer from a similar pattern.
Resource lookup maybe?

Also, we're processing the mentioned queue on the authorization thread.
Should we move that processing back to the channel's event thread?
Can we do that in netty?

Where is this netty-4 threading model that was spoken of?